modified

+++ sqroot.py
@@ -1,5 +1,5 @@
 #find the square root of any number
 
-x=input("Enter your number: ")
-a = int(x) ** 0.5
+x=float(input("Enter your number: "))
+a = x ** 0.5
 print("The square root of the number is: ", a)
